requirements.txt
asgiref==3.8.1
Django==5.0.4
django-countries==7.6.1
django-phonenumber-field==7.3.0
django-utils-six==2.0
phonenumberslite==8.13.35
pillow==10.3.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
